muda tudo pra double

+++ csv_parser.cpp
@@ -6,12 +6,12 @@
 #include <string>
 #include <iostream>
 
-std::vector<std::pair<std::string, std::pair<float, float>>> csv_parser() {
+std::vector<std::pair<std::string, std::pair<double, double>>> csv_parser() {
     std::string file_name = "BR_Municipios_2025.csv";
     std::ifstream file(file_name);
     std::string line;
 
-    std::vector<std::pair<std::string, std::pair<float, float>>> cities;
+    std::vector<std::pair<std::string, std::pair<double, double>>> cities;
     std::string cur, name, latitude, longitude;
 
     while (std::getline(file, line)) {
@@ -29,7 +29,7 @@ std::vector<std::pair<std::string, std::pair<float, float>>> csv_parser() {
         longitude = cur;
 
         if (latitude[0] != 'L')  // A primeira linha isso vai ser "LATITUDE"
-            cities.emplace_back(name, std::make_pair(std::stof(latitude), std::stod(longitude)));
+            cities.emplace_back(name, std::make_pair(std::stod(latitude), std::stod(longitude)));
     }
 
     file.close();
